requirements.txt
# =============================================================================
# Mindspace — Text Sentiment API (main.py, LightGBM, port 9000)
# =============================================================================

# ── API framework ─────────────────────────────────────────────────────────────
fastapi==0.134.0
uvicorn[standard]==0.41.0
pydantic==2.12.5

# ── ML inference ──────────────────────────────────────────────────────────────
numpy==2.4.2
pandas==2.3.3
scikit-learn==1.8.0
lightgbm==4.6.0
joblib==1.5.3

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.2.1
requests==2.32.3
